+++ cogstat/cogstat_gui.py
@@ -79,7 +79,7 @@ class StatMainWindow(QtWidgets.QMainWindow):
 
         # Only for testing
 #        self.open_file('cogstat/sample_data/example_data.csv'); #self.compare_groups()
-#        self.open_file('cogstat/test/data/diffusion.csv')
+#        self.open_file('cogstat/test/data/VA_test.csv')
 #        self.open_clipboard()
 #        self.print_data()
 #        self.explore_variable(['X'])
@@ -97,6 +97,7 @@ class StatMainWindow(QtWidgets.QMainWindow):
 #        self.compare_groups(['slope'], ['group'],  ['slope_SE'], 25)
 #        self.compare_groups(['A'], ['G', 'H'])
 #        self.compare_groups(['X'], ['TIME', 'CONDITION'])
+#        self.compare_groups(['dep_nom'], ['g0', 'g1', 'g2', 'g3'])
 #        self.save_result_as()
 #        self.save_result_as(filename='CogStat analysis result.pdf')