+++ changelog.md
@@ -6,6 +6,7 @@ Upcoming release
 
 ## New features
 - Statistical analysis
+    - Any number of grouping variables in Compare groups
     - Contingency table improvements (in Explore relation of variable pairs, Compare repeated measures variables, Compare groups)
         - Display margins
         - Display percentage